moistmaster/requirements/base.txt
pytz==2018.4  # https://github.com/stub42/pytz

# Django
# ------------------------------------------------------------------------------
django==2.0.6  # pyup: < 2.1  # https://www.djangoproject.com/
django-compressor==2.2  # https://github.com/django-compressor/django-compressor
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-crispy-forms==1.7.2
django-localflavor==2.2

# Robosquirt libs
# ------------------------------------------------------------------------------
ciso8601==2.0.1
pyzmq==17.0.0
tornado==5.0.2
zmq==0.0.0
requests==2.19.1
